feat/TKAPAP-10-SIFACTORY

+++ SI-FACTORY/src/main/java/apap/group/assignment/SIFACTORY/repository/DeliveryDB.java
@@ -4,5 +4,5 @@ import apap.group.assignment.SIFACTORY.model.DeliveryModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DeliveryDB extends JpaRepository<DeliveryModel, Long> {
-    DeliveryModel findByIdDelivery(Integer id);
+    DeliveryModel findByIdDelivery(Long id);
 }
