+++ SI-FACTORY/src/main/java/apap/group/assignment/SIFACTORY/repository/RequestUpdateItemDB.java
@@ -4,5 +4,5 @@ import apap.group.assignment.SIFACTORY.model.RequestUpdateItemModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RequestUpdateItemDB extends JpaRepository<RequestUpdateItemModel, Long> {
-    RequestUpdateItemModel findByIdRequestUpdateItem(Integer id);
+    RequestUpdateItemModel findByIdRequestUpdateItem(Long id);
 }
